requirements.txt
beautifulsoup4==4.12.0
Django==4.1.7
djangorestframework==3.14.0
mysql_connector_repackaged==0.3.1
requests==2.28.2
selenium==4.8.3
webdriver_manager==3.8.5
